Moved tests to separate file
